Convert the last pages to app router
