+++ app/package.json
@@ -66,7 +66,7 @@
     "lucide-react": "^0.404.0",
     "mysql2": "^3.10.1",
     "nanoid": "^4.0.2",
-    "next": "14.2.4",
+    "next": "14.2.5",
     "next-axiom": "1.1.1",
     "node-html-markdown": "^1.3.0",
     "openai": "^4.19.1",
