+++ app/src/app/jutsus/page.tsx
@@ -1,0 +1,256 @@
+"use client";
+
+import { useState } from "react";
+import { Trash2 } from "lucide-react";
+import ItemWithEffects from "@/layout/ItemWithEffects";
+import ContentBox from "@/layout/ContentBox";
+import Modal from "@/layout/Modal";
+import Loader from "@/layout/Loader";
+import LoadoutSelector from "@/layout/LoadoutSelector";
+import Confirm from "@/layout/Confirm";
+import { SquareChevronRight, SquareChevronLeft } from "lucide-react";
+import { Button } from "@/components/ui/button";
+import { ActionSelector } from "@/layout/CombatActions";
+import { calcJutsuEquipLimit, calcForgetReturn } from "@/libs/train";
+import { useRequiredUserData } from "@/utils/UserContext";
+import { api } from "@/utils/api";
+import { showMutationToast } from "@/libs/toast";
+import { JUTSU_XP_TO_LEVEL } from "@/drizzle/constants";
+import type { Jutsu, UserJutsu } from "@/drizzle/schema";
+
+export default function MyJutsu() {
+  // tRPC utility
+  const utils = api.useUtils();
+
+  // Settings
+  const now = new Date();
+  const { data: userData } = useRequiredUserData();
+  const [isOpen, setIsOpen] = useState<boolean>(false);
+  const [userjutsu, setUserJutsu] = useState<(Jutsu & UserJutsu) | undefined>(
+    undefined,
+  );
+
+  // User Jutsus & items
+  const { data: userJutsus, isFetching: l1 } = api.jutsu.getUserJutsus.useQuery(
+    undefined,
+    { staleTime: Infinity },
+  );
+  const { data: userItems, isFetching: l2 } = api.item.getUserItems.useQuery(
+    undefined,
+    { staleTime: Infinity },
+  );
+
+  const userJutsuCounts = userJutsus?.map((userJutsu) => {
+    return {
+      id: userJutsu.id,
+      quantity:
+        userJutsu.finishTraining && userJutsu.finishTraining > now
+          ? userJutsu.level - 1
+          : userJutsu.level,
+    };
+  });
+
+  // Mutations
+  const { mutate: equip, isPending: isEquipping } = api.jutsu.toggleEquip.useMutation({
+    onSuccess: async (data) => {
+      showMutationToast(data);
+      if (data.success) {
+        await utils.jutsu.getUserJutsus.invalidate();
+      }
+    },
+    onSettled: () => {
+      document.body.style.cursor = "default";
+      setIsOpen(false);
+      setUserJutsu(undefined);
+    },
+  });
+
+  const { mutate: forget, isPending: isForgetting } = api.jutsu.forget.useMutation({
+    onSuccess: async (data) => {
+      showMutationToast(data);
+      if (data.success) {
+        await utils.jutsu.getUserJutsus.invalidate();
+      }
+    },
+    onSettled: () => {
+      document.body.style.cursor = "default";
+      setIsOpen(false);
+      setUserJutsu(undefined);
+    },
+  });
+
+  const { mutate: updateOrder } = api.jutsu.updateUserJutsuOrder.useMutation({
+    onSuccess: async (data) => {
+      showMutationToast(data);
+      if (data.success) {
+        await utils.profile.getUser.invalidate();
+      }
+    },
+  });
+
+  const isPending = isEquipping || isForgetting;
+  const isFetching = l1 || l2;
+
+  // Collapse UserItem and Item
+  const allJutsu = userJutsus?.map((userjutsu) => {
+    let warning = "";
+    if (userjutsu.jutsu.jutsuWeapon !== "NONE") {
+      const equippedItem = userItems?.find(
+        (useritem) =>
+          useritem.item.weaponType === userjutsu.jutsu.jutsuWeapon &&
+          useritem.equipped !== "NONE",
+      );
+      if (!equippedItem) {
+        warning = `No ${userjutsu.jutsu.jutsuWeapon.toLowerCase()} weapon equipped.`;
+      }
+    }
+    return {
+      ...userjutsu.jutsu,
+      ...userjutsu,
+      highlight: userjutsu.equipped ? true : false,
+      warning: warning,
+    };
+  });
+
+  // Sort if we have a loadout
+  if (userData?.loadout?.jutsuIds && allJutsu) {
+    allJutsu.sort((a, b) => {
+      const aIndex = userData?.loadout?.jutsuIds.indexOf(a.jutsuId) ?? -1;
+      const bIndex = userData?.loadout?.jutsuIds.indexOf(b.jutsuId) ?? -1;
+      if (aIndex === -1 && bIndex === -1) return 0;
+      if (aIndex === -1) return 1;
+      if (bIndex === -1) return -1;
+      return aIndex - bIndex;
+    });
+  }
+
+  // Derived calculations
+  const curEquip = userJutsus?.filter((j) => j.equipped).length;
+  const maxEquip = userData && calcJutsuEquipLimit(userData);
+  const canEquip = curEquip !== undefined && maxEquip && curEquip < maxEquip;
+  const subtitle =
+    curEquip && maxEquip
+      ? `Equipped ${curEquip}/${maxEquip}`
+      : "Jutsus you want to use in combat";
+
+  // Ryo from forgetting
+  const forgetRyo = userjutsu
+    ? calcForgetReturn(userjutsu, userjutsu.level).toFixed(0)
+    : 0;
+
+  // Loaders
+  if (!userData) {
+    return <Loader explanation="Loading userdata" />;
+  } else if (isFetching) {
+    return <Loader explanation="Loading jutsus" />;
+  }
+
+  return (
+    <ContentBox
+      title="Jutsu Management"
+      subtitle={subtitle}
+      topRightContent={!isOpen && <LoadoutSelector />}
+    >
+      <ActionSelector
+        items={allJutsu}
+        counts={userJutsuCounts}
+        labelSingles={true}
+        onClick={(id) => {
+          setUserJutsu(allJutsu?.find((jutsu) => jutsu.id === id));
+          setIsOpen(true);
+        }}
+        showBgColor={false}
+        showLabels={true}
+        emptyText="You have not learned any jutsu. Go to the training grounds in your village to learn some."
+      />
+      {isOpen && userData && userjutsu && (
+        <Modal
+          title="Edit Jutsu"
+          proceed_label={
+            !isEquipping
+              ? userjutsu.equipped
+                ? "Unequip"
+                : canEquip
+                  ? "Equip"
+                  : "Unequip other first"
+              : undefined
+          }
+          setIsOpen={setIsOpen}
+          isValid={false}
+          onAccept={() => {
+            if (canEquip || userjutsu.equipped) {
+              equip({ userJutsuId: userjutsu.id });
+            } else {
+              setIsOpen(false);
+            }
+          }}
+          confirmClassName={
+            canEquip
+              ? "bg-blue-600 text-white hover:bg-blue-700"
+              : "bg-red-600 text-white hover:bg-red-700"
+          }
+        >
+          <p>- You have {userData.money} ryo in your pocket</p>
+          <p className="pb-3">
+            - Need {JUTSU_XP_TO_LEVEL - userjutsu.experience} XP more to level
+          </p>
+          {!isPending && (
+            <>
+              <ItemWithEffects
+                item={userjutsu}
+                key={userjutsu.id}
+                showStatistic="jutsu"
+              />
+              <div className="flex flex-row gap-3 items-center">
+                {userData.loadout?.jutsuIds.includes(userjutsu.jutsuId) && (
+                  <>
+                    <SquareChevronLeft
+                      className="h-8 w-8 hover:text-orange-300 hover:cursor-pointer"
+                      onClick={() =>
+                        updateOrder({
+                          jutsuId: userjutsu.jutsuId,
+                          loadoutId: userData?.jutsuLoadout ?? "",
+                          moveForward: false,
+                        })
+                      }
+                    />
+                    <p>Order</p>
+                    <SquareChevronRight
+                      className="h-8 w-8 hover:text-orange-300 hover:cursor-pointer"
+                      onClick={() =>
+                        updateOrder({
+                          jutsuId: userjutsu.jutsuId,
+                          loadoutId: userData?.jutsuLoadout ?? "",
+                          moveForward: true,
+                        })
+                      }
+                    />
+                  </>
+                )}
+
+                <div className="grow"></div>
+                <Confirm
+                  title="Forget Jutsu"
+                  button={
+                    <Button id="return" variant="destructive">
+                      <Trash2 className="h-6 w-6 mr-2" />
+                      Forget [${forgetRyo} ryo]
+                    </Button>
+                  }
+                  onAccept={(e) => {
+                    e.preventDefault();
+                    forget({ id: userjutsu.id });
+                  }}
+                >
+                  <p>Confirm to forget this jutsu and get back {forgetRyo} ryo.</p>
+                </Confirm>
+              </div>
+            </>
+          )}
+          {isPending && <Loader explanation={`Processing ${userjutsu.name}`} />}
+        </Modal>
+      )}
+      {isPending && <Loader explanation="Loading Jutsu" />}
+    </ContentBox>
+  );
+}
